refactor(database): change data types of columns

+++ crates/models/database/src/daily_user_activity.rs
@@ -16,10 +16,10 @@ pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
     #[graphql(skip)]
     pub user_id: String,
-    pub total_counts: i64,
-    pub review_counts: i64,
-    pub workout_counts: i64,
-    pub measurement_counts: i64,
+    pub total_counts: i32,
+    pub review_counts: i32,
+    pub workout_counts: i32,
+    pub measurement_counts: i32,
     pub total_duration: Decimal,
     #[sea_orm(column_type = "Json")]
     #[graphql(skip)]
